feat(crawler): replace createWebSourceConstructor with webSourceConstructor and update data sources
